more review fixes
