+++ src/core/app/features.ts
@@ -1,23 +1,25 @@
 import { appConfig } from '~core/app_config';
-import { apiClient } from '~core/apiClientInstance';
-import type { AppFeatureType, BackendFeature } from '~core/auth/types';
+import { featureFlagsAtom } from '~core/shared_state';
+import type { BackendFeature } from '~core/auth/types';
+import type { ApiClient } from '~core/api_client';
 
-const effectiveFeatures = {};
-
-// intersection of both, user can use his feature if app has it
-function mergeFeatures(global, custom, user) {
-  const res = { ...global, ...custom, ...user };
-  return res;
-}
-
-export function hasFeature(f: AppFeatureType) {
-  return effectiveFeatures[f];
-}
-
-export async function getFeatures(useAuth: boolean) {
-  const featuresResponse = await apiClient.get<BackendFeature[]>(
+export function loadFeatures(
+  apiClient: ApiClient,
+  useAuth: boolean,
+): Promise<BackendFeature[] | null> {
+  const featuresResponse = apiClient.get<BackendFeature[]>(
     `/features`,
     { appId: appConfig.id },
     useAuth,
   );
+  return featuresResponse;
+}
+
+export function setFeatures(value: BackendFeature[] | null) {
+  // FIXME: investigate proper app and user feature merge
+  const newFeatures = {};
+  value?.forEach((ft) => {
+    newFeatures[ft.name] = true;
+  });
+  featureFlagsAtom.set.dispatch(newFeatures);
 }
