+++ src/core/shared_state/currentApplication.ts
@@ -1,5 +1,4 @@
 import { createPrimitiveAtom } from '~utils/atoms/createPrimitives';
-// import { appConfig } from '~core/app_config';
 
 export const currentApplicationAtom = createPrimitiveAtom<null | string>(
   '',
